# API/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Banka is a light-weight core banking application that powers banking operations like account creation, customer deposit and withdrawals.",
    "version": "1.0.0",
    "title": "Banka API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "tejumoladavid@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:7888",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/auth/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login with existing credentials",
        "description": "",
        "operationId": "loginUser",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object credentials that needs to sign in",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "A User Profile Cannot be Found!"
          },
          "422": {
            "description": "Validation error | Bad request"
          },
          "200": {
            "description": "Successfully login a user"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
      "bearerAuth": {
        "type": "apiKey",
        "scheme": "bearer",
        "description": "JWT token is received after registering/login, input format: \"Bearer \\<token\\>\"",
        "name": "x-access-token",
        "in": "header",
        "bearerFormat": "JWT"
    }
  },
  "definitions": {
...
    "Login": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "tejumolatimi@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "password"
        }
      },
      "xml": {
        "name": "Login"
      }
    },
...
  },
...
}